Move a lot of terminal into browser
Fixes #79425
