+++ src/vs/workbench/contrib/terminal/electron-browser/windowsShellHelper.ts
@@ -5,10 +5,11 @@
 
 import * as platform from 'vs/base/common/platform';
 import { Emitter, Event } from 'vs/base/common/event';
-import { ITerminalInstance, IWindowsShellHelper, TitleEventSource } from 'vs/workbench/contrib/terminal/common/terminal';
+import { IWindowsShellHelper, TitleEventSource } from 'vs/workbench/contrib/terminal/common/terminal';
 import { Terminal as XTermTerminal } from 'xterm';
 import * as WindowsProcessTreeType from 'windows-process-tree';
 import { Disposable } from 'vs/base/common/lifecycle';
+import { ITerminalInstance } from 'vs/workbench/contrib/terminal/browser/terminal';
 
 const SHELL_EXECUTABLES = [
 	'cmd.exe',
